added a noteManger.cs to the file
